requirements.txt
discord.py==2.4.0
PyNaCl==1.5.0
PyQt6==6.7.1
keyboard==0.13.5
